moving brower mock setup to test folder
