+++ src/Frontend/package.json
@@ -63,5 +63,10 @@
     "vite-plugin-checker": "^0.6.4",
     "vitest": "^1.6.0",
     "vue-tsc": "^1.8.25"
+  },
+  "msw": {
+    "workerDirectory": [
+      "public"
+    ]
   }
 }
